BugFixes: favorite cover in LS was null; repaired logic for favorite photos; after removing a few photos, others marked as unfavorite(only on frontend); the request is no longer being sent to album controller with Id = 0; Favorite cover set as last.

+++ WebClient/lama/src/app/components/view-album-module/view-album/view-album.component.ts
@@ -52,9 +52,9 @@ export class ViewAlbumComponent implements OnInit, DoCheck {
   ngOnInit() {
     const userId: number = parseInt(localStorage.getItem('userId'), 10);
     this.selectedPhotos = [];
-    if (this.loading === false && this.AlbumId !== 0) {
+    if (this.loading == false && this.AlbumId != 0) {
       this.albumService.getAlbum(this.AlbumId).subscribe( x => {this.album = x.body; });
-    } else if (this.AlbumId === 0) {
+    } else if (this.AlbumId == 0) {
             this.favoriteService.getFavoritesPhotos(userId)
            .subscribe(data => {
             this.album.photoAlbums = data;
@@ -87,13 +87,11 @@ export class ViewAlbumComponent implements OnInit, DoCheck {
 
   ngDoCheck() {
     this.isAtLeastOnePhotoSelected = this.selectedPhotos.length > 0;
-    if (this.album.photoAlbums !== undefined) {
-      if (this.album.photoAlbums.length === 0) {
-        if (this.album.id === 0) {
-          localStorage.removeItem('favoriteCover');
-        } else {
-          this.albumService.removeAlbumCover(this.album.id).subscribe(x => x);
-        }
+    if (this.album.photoAlbums != undefined && this.album.photoAlbums.length == 0) {
+      if (this.album.id == 0) {
+        localStorage.removeItem('favoriteCover');
+      } else {
+        this.albumService.removeAlbumCover(this.album.id).subscribe(x => x);
       }
     }
   }
@@ -119,19 +117,21 @@ export class ViewAlbumComponent implements OnInit, DoCheck {
     this.album.photoAlbums.forEach(element => {
       ids.push(element.id);
     });
-    if (this.AlbumId === 0) {
+    if (this.AlbumId == 0) {
       this.selectedPhotos.forEach(item => {
         this.favoriteService.deleteFavorite(parseInt(localStorage.getItem('userId'), 10), item.id).subscribe(() => {
-          this.favorites = new Set<number>();
+          this.favorites.delete(item.id);
         });
       });
     }
-    this.selectedPhotos = [];
-    this.albumService.updateAlbum({
-      title: this.album.title,
-      id: this.album.id,
-      photoIds: ids
-    });
+    else {
+      this.selectedPhotos = [];
+      this.albumService.updateAlbum({
+        title: this.album.title,
+        id: this.album.id,
+        photoIds: ids
+      });
+    }
   }
 
   public downloadImages() {
